Fix sign mismatch warning in 32-bit builds

+++ src/render/d3d11/tex_mgr/d3d11_tex_mgr.cpp
@@ -2504,11 +2504,11 @@ SK_D3D11_TextureIsCachedEx (ID3D11Texture2D* pTex, bool touch = false)
     bool use_l3 =
       config.textures.d3d11.use_l3_hash;
 
-    if (( use_l3 && tex_desc.Format    < textures->HashMap_Fmt.size () &&
-                    tex_desc.MipLevels < textures->HashMap_Fmt [tex_desc.Format].map.size () &&
-                                         textures->HashMap_Fmt [tex_desc.Format].map [tex_desc.MipLevels].contains (pTex)) ||
-        (!use_l3 && tex_desc.MipLevels < textures->HashMap_2D.size ()  &&
-                                         textures->HashMap_2D [tex_desc.MipLevels].contains (pTex)))
+    if (( use_l3 && (size_t)tex_desc.Format    < textures->HashMap_Fmt.size () &&
+                            tex_desc.MipLevels < textures->HashMap_Fmt [tex_desc.Format].map.size () &&
+                                                 textures->HashMap_Fmt [tex_desc.Format].map [tex_desc.MipLevels].contains (pTex)) ||
+        (!use_l3 &&         tex_desc.MipLevels < textures->HashMap_2D.size ()  &&
+                                                 textures->HashMap_2D [tex_desc.MipLevels].contains (pTex)))
     {
       if (touch && (! SK_D3D11_IsTexInjectThread ()))
       {
